fastapi_app/requirements.txt
python-dotenv==1.0.1
langchain-openai==0.2.14
langchain==0.3.13
langgraph==0.2.60
pydantic==2.10.1
fastapi[all]==0.115.6
azure-search-documents==11.6.0b8 
azure-identity==1.19.0